requirements.txt
streamlit
chromadb
sentence-transformers
transformers
torch
faiss-cpu
requests
beautifulsoup4
python-dotenv
pyyaml
pydantic
sqlalchemy
uvicorn
fastapi
pytest
pytest-cov
pyttsx3
pydub
werkzeug
huggingface_hub
accelerate
groq
streamlit-mic-recorder
gTTS
streamlit-audiorecorder
SpeechRecognition
pydub
streamlit-webrtc
SpeechRecognition
av
pyaudio
streamlit-audio-recorder

